requirements.txt
flask
flask-sqlalchemy
flask-wtf
stripe
